requirements.txt
# Basic requirements
PyQt6==6.4.0
python-dotenv==0.19.0

# Google Gemini API
google-generativeai==0.3.1 